Conflicts:
	config.codekit

+++ craft/app/services/AssetTransformsService.php
@@ -37,6 +37,11 @@ class AssetTransformsService extends BaseApplicationComponent
 	 */
 	private $_eagerLoadedTransformIndexes;
 
+	/**
+	 * @var
+	 */
+	private $_activeTransformIndexModel;
+
 	// Public Methods
 	// =========================================================================
 
@@ -418,7 +423,7 @@ class AssetTransformsService extends BaseApplicationComponent
 			{
 				// Wait a second!
 				sleep(1);
-				ini_set('max_execution_time', 120);
+				craft()->config->maxPowerCaptain();
 
 				$index = $this->getTransformIndexModelById($index->id);
 
@@ -434,7 +439,6 @@ class AssetTransformsService extends BaseApplicationComponent
 					}
 					else
 					{
-						$index->dateUpdated = new DateTime();
 						$this->storeTransformIndexData($index);
 						break;
 					}
@@ -1050,11 +1054,14 @@ class AssetTransformsService extends BaseApplicationComponent
 	{
 		$thumbFolders = IOHelper::getFolderContents(craft()->path->getAssetsThumbsPath());
 
-		foreach ($thumbFolders as $folder)
+		if ($thumbFolders)
 		{
-			if (is_dir($folder))
+			foreach ($thumbFolders as $folder)
 			{
-				IOHelper::deleteFile($folder.'/'.$file->id.'.'.$this->_getThumbExtension($file));
+				if (is_dir($folder))
+				{
+					IOHelper::deleteFile($folder.'/'.$file->id.'.'.$this->_getThumbExtension($file));
+				}
 			}
 		}
 	}
@@ -1092,6 +1099,22 @@ class AssetTransformsService extends BaseApplicationComponent
 			->queryAll();
 
 		return AssetTransformIndexModel::populateModels($records);
+	}
+
+	/**
+	 * @return mixed
+	 */
+	public function getActiveTransformIndexModel()
+	{
+		return $this->_activeTransformIndexModel;
+	}
+
+	/**
+	 * @param AssetTransformIndexModel $index
+	 */
+	public function setActiveTransformIndexModel(AssetTransformIndexModel $index)
+	{
+		$this->_activeTransformIndexModel = $index;
 	}
 
 	// Private Methods
@@ -1208,6 +1231,9 @@ class AssetTransformsService extends BaseApplicationComponent
 		{
 			$image->setQuality($quality);
 		}
+
+		// Save this for Image to use if needed.
+		$this->setActiveTransformIndexModel($index);
 
 		switch ($transform->mode)
 		{
